+++ core/src/main/scala/dx/core/ir/ParameterLink.scala
@@ -296,4 +296,8 @@ case class ParameterLinkDeserializer(dxFileDescCache: DxFileDescCache, dxApi: Dx
                                    dxApi,
                                    Some(dxFileDescCache))
   }
+
+  def updateWithCache(dxFileDescCache: DxFileDescCache): ParameterLinkDeserializer = {
+    ParameterLinkDeserializer(dxFileDescCache)
+  }
 }