+++ RELEASE_NOTES.md
@@ -1,6 +1,8 @@
 # Release Notes
 
 ## in develop
+* WDL: optimizations to reduce unnecessary `file-xxx/describe` API calls upon (de)localization of input/output files. 
+Now all files are described in bulk.
 * Fixes for the `-pretty` flag in the `describe` command
 * WDL: fix for the tasks wrapped in a frag with optional inputs. Those inputs get forcibly evaluated to explicit `null` 
 and cause inconsistencies with applet inputs. This is fixed and all implicit `null` (`None`) values of optionals are remaining implicit.